+++ src/Controller/MesuresController.php
@@ -107,22 +107,29 @@ class MesuresController extends AppController
         	
         	//debug($this->request->data);die();
         	$d = $this->request->data;
-        	$nomFichier = $this->Utilitaire->replaceCaracterespeciaux($d['file']['name']);  
-        	$destination = DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$nomFichier;
-        	move_uploaded_file($d['file']['tmp_name'], $destination);
-        	
-        	$mesure->id = null;
-        	$mesure->demarche_id = $d['demarche_id'];
-        	$mesure->name = $d['name'];
-        	$mesure->resultat = $d['resultat'];
-        	$mesure->file = $nomFichier;
-        	 
-            if ($this->Mesures->save($mesure)) {
-                $this->Flash->success('La mesure a bien été sauvegardée.');
-                return $this->redirect(['action' => 'index']);
-            } else {
-                $this->Flash->error('Erreur dans la sauvegarde de la mesure.');
-            }
+
+        	//Quand la taille du fichier est trop importante
+        	// c'est à dire les paramètres du php.ini (upload_max_filesize et post_max_size) => formulaire vide
+        	if(empty($d)) {
+        		$this->Flash->error('La taille du fichier dépasse la limite des 10 Mo.');
+        	} else {
+        		$nomFichier = $this->Utilitaire->replaceCaracterespeciaux($d['file']['name']);  
+	        	$destination = DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$nomFichier;
+	        	move_uploaded_file($d['file']['tmp_name'], $destination);
+	        	
+	        	$mesure->id = null;
+	        	$mesure->demarche_id = $d['demarche_id'];
+	        	$mesure->name = $d['name'];
+	        	$mesure->resultat = $d['resultat'];
+	        	$mesure->file = $nomFichier;
+	        	 
+	            if ($this->Mesures->save($mesure)) {
+	                $this->Flash->success('La mesure a bien été sauvegardée.');
+	                return $this->redirect(['action' => 'index']);
+	            } else {
+	                $this->Flash->error('Erreur dans la sauvegarde de la mesure.');
+	            }
+        	}
         }
         $this->set(compact('mesure','demarche_id'));
         $this->set('_serialize', ['mesure']);
@@ -154,25 +161,39 @@ class MesuresController extends AppController
         	 
         	//Cas d'un nouveau fichier : CRM et Culture securite
         	if(isset($d['file']) && $d['file']['tmp_name'] != '') {
-        		// Il s'agit d'un nouveau fichier
-        		//Vérification de la présence
-        		if(file_exists(DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$d['file']['name'])) {
-        			$boolSupp = unlink(DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$d['file']['name']);
+
+        		//Quand la taille du fichier est trop importante
+        		// c'est à dire les paramètres du php.ini (upload_max_filesize et post_max_size) => formulaire vide
+        		if(empty($d)) {
+        			$this->Flash->error('La taille du fichier dépasse la limite des 10 Mo.');
+        		} else {
+	        		// Il s'agit d'un nouveau fichier
+	        		//Vérification de la présence
+	        		if(file_exists(DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$d['file']['name'])) {
+	        			$boolSupp = unlink(DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$d['file']['name']);
+	        		}
+	        		//Deplacement du nouveau
+	        		$nomFichier = $this->Utilitaire->replaceCaracterespeciaux($d['file']['name']);  
+	        		$destination = DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$nomFichier;
+	        		move_uploaded_file($d['file']['tmp_name'], $destination);
         		}
-        		//Deplacement du nouveau
-        		$nomFichier = $this->Utilitaire->replaceCaracterespeciaux($d['file']['name']);  
-        		$destination = DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$nomFichier;
-        		move_uploaded_file($d['file']['tmp_name'], $destination);
         	} else if(isset($d['file_new']) && $d['file_new']['tmp_name'] != '') {
-        		//Cas d'une modification
-        		//Suppression de l'ancien
-        		if(file_exists(DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$mesure->file) && strlen($mesure->file)>0) {
-        			$boolSupp = unlink(DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$mesure->file);
+
+        		//Quand la taille du fichier est trop importante
+        		// c'est à dire les paramètres du php.ini (upload_max_filesize et post_max_size) => formulaire vide
+        		if(empty($d)) {
+        			$this->Flash->error('La taille du fichier dépasse la limite des 10 Mo.');
+        		} else {
+	        		//Cas d'une modification
+	        		//Suppression de l'ancien
+	        		if(file_exists(DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$mesure->file) && strlen($mesure->file)>0) {
+	        			$boolSupp = unlink(DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$mesure->file);
+	        		}
+	        		//Deplacement du nouveau
+	        		$nomFichier = $this->Utilitaire->replaceCaracterespeciaux($d['file_new']['name']); 
+	        		$destination = DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$nomFichier;
+	        		move_uploaded_file($d['file_new']['tmp_name'], $destination);
         		}
-        		//Deplacement du nouveau
-        		$nomFichier = $this->Utilitaire->replaceCaracterespeciaux($d['file_new']['name']); 
-        		$destination = DATA.'userDocument'.DS.$session->read('Auth.User.username').DS.$nomFichier;
-        		move_uploaded_file($d['file_new']['tmp_name'], $destination);
         	} else {
         		//Pas de nouveau fichier et pas de modification de fichier : modification des autres champs textes du formulaire
         		$nomFichier = $mesure->file ;
